requirements.txt
django
djangorestframework
django_filter
factory_boy
